Pokemon_App/backend/requirements.txt
fastapi
uvicorn
pandas
sqlalchemy
pymysql
python-dotenv
cryptography
pydantic
httpx